// types/superadmin.d.ts
export interface FiltrosLogs {
    admin_id?: string;
    accion?: string;
    entidad_tipo?: 'empresa' | 'usuario' | 'suscripcion' | 'pago';
    fecha_desde?: string;
    fecha_hasta?: string;
}